Accept cookiecutter arguments on the command line for update
This commit adds support for providing cookiecutter arguments on the
command line for `cruft update`. If any arguments are provided on the
command line `cruft update` will always render the template regardless
of whether the user has requested to update to a new template version or
not.

Ideally this change will be applied after
https://github.com/cruft/cruft/pull/159/ as it's been developed on top
of that PR originally.